Add files via upload
Now the outer-product and dot-product  can be computed using the * operator. The correct computation depends on the orientation of the vectors that are being multiplied. Also we can print the result of this operation. Printing the result of a multiplication is tough, because multiplication can return either the dot product (a float) or a (cb::Matrix). Handling the std::variant<float, cb::Matrix> was tough.