Add files via upload
Rewrote the Vector class again. I still need to add all its mathematical function back. Now vectors can be created with any type. Obviously this is a linear algebra library, so filling the vectors with strings is probably not a good idea. Moreover, I have adhered to the rule of 5. Thus Vectors should not waste resources.